Move all auto completion code to a separate file
